+++ WebContent/common/mypage_seller_menu.jsp
@@ -8,7 +8,7 @@
 
 		<li><a href="${CONTEXT_PATH}/member/mypageController?action=sellerInfoForm">내 정보 조회</a></li>
 		<li><a href="${CONTEXT_PATH}/order/orderController?action=sellerOrderList">판매주문 목록</a></li>
-		<li><a href="${CONTEXT_PATH}/member/mypageController?action=itemaddForm">상품 등록</a></li>
+		<li><a href="${CONTEXT_PATH}/item/itemController?action=itemEnrollForm">상품 등록</a></li>
 		<li><a href="${CONTEXT_PATH}/member/mypageController?action=removeMemberForm">회원 탈퇴</a></li>
 		<li><a href="${CONTEXT_PATH}/item/itemController?action=myitemList">등록한 상품조회</a></li>
